Switch sim to terminate on main thread termination
Originally the simulator followed the behaviour that it would terminate
when there were no more runnable threads (and no remaining timers).

This does not match GHC's behaviour which terminates when the main
thread terminates. So we switch the sim to this behaviour.

In addition this makes it easier to return the result of the main
thread, so we do that because it's very convenient for tests.

We also return the list of running threads on termination of the main
thread. This makes it possible to check for a clean shutdown if that is
of interest. For example if one is simulating a tree of threads and one
wants to check all threads are shutdown cleanly.