Add an easy to use runSim
runSim :: forall a. (forall s. SimM s a) -> Either Failure a
If one does not care about the trace, this makes getting the final
result very easy.

Als add runSimStrictShutdown which checks that there are no more
runnable or blocked threads left when the main thread terminates. This
should be useful for testing thread cleanup code.