Track thread running/blocked state more precisely
Add threadBlocked :: Bool to Thread. Extend the invariant to make sure
this is consistent with the runqueue.

To preserve the invariant more consistently, use a new unblockThreads
helper to update the threads state when the runqueue is updated.

Also rely on the new flag in various places rather than checking the
runqueue.