Use a fixed AsyncCancelled exception type
Rather than allowing it to be different for different implementations.
This eliminates the otherwise annoying isCancel method of the MonadAsync
class. Also re-export the type so it can be used more easily.

This also makes it possible to throw the concrete exception type, which
is what the next patch needs this for.