Add getCurrentTime :: m UTCTime to MonadTime and sim
This gives us two clocks, a monotonic clock and a wall clock. In IO the
wall clock is the normal getCurrentTime from the time package.

In the simulator there is a new notion of clock domain. Each clock
domain has a current wall clock time in UTCTime. Each thread belongs
to a clock domain.

In the simulator only, the wall clock time can also be set. This does
not affect the monotonic clock.

In the simulator only, a thread can put itself into a new clock domain
which is not shared with other existing threads. This is intended to
allow simulation of code running on different hosts that each have their
own local wall clocks that are not necessarily synchronised.

This is intended to be used for testing the consensus implementation
when there is clock skew between nodes.