io-sim: extend tvar tracing with string values
Sometimes it is useful to trace string values when inspecting 'TVar'
values (see next commit in which we trace when a 'TMVar' is taken
/ released).  For this purpose we provide a 'TraceValue' record and some
useful pattern synonyms, which allow to log dynamic and / or string
values.  IOSim will log Dynamic values with 'EventDynamic', while string
values will be logged with 'EventSay'.  For 'IO' both are ignored, as
the 'traceTVar' callback never fires.