Generalise OrElseLeft/Right frame
OrElseLeftFrame and OrElseRightFrame represent a control structure that
has a alternative branch that is executed when `retry` is applied. A
`catch` has a similar execution model when a `throw` is applied.

The control frame is generalised to BranchFrame that can hold an
alternative statement. If the execution context is `left` side of the
branch then the BranchFrame contains `right` statement. When we are
executing in the `right` context, the branch frame contains an empty
statement.