MonadTimer & MonadDelay in si-timers
`si-timers` also provides `MonadTimer` and `MonadDelay`.  They use the
same name as one in `io-classes` as it's not expected to depend on both
in one module.

The `si-timers` provides `defaultRegisterDelayCancellable` which is used
for two purposes:

* provide `IO` instance
* test the `IO` instance in `IOSim`

This patch also reverts a change in one of the previous patches.
`IOSim` is using `DiffTime` for its `NewTimeout` primitive operation.
The `UpdateTimeout` is removed as it is not used.