MVar: fix missing state updates in {try}takeMVarDefault
In the default implementation of MVar (used in the SimM, not IO), there
were two cases where the mvar state was not being correctly updated. In
both takeMVarDefault and tryTakeMVarDefault, in the case of a full mvar
with an empty wait queue, the mvar state was not being updated.

This is likely to be the cause of deadlocks observed in some simulations
using MVars.