MVar: fix isEmptyMVarDefault invariant violation
The existing code assumed for some reason that it is impossible to have
any threads waiting on an empty mvar when someone is testing it to see
if it is empty. This is false, it's entirely possible, and doesn't
affect the question of whether the mvar is empty or not.