Do not require the MVar type family to be injective
We don't require this for TVar or other classes. We only currently use
injective type families for the monad (like STM) not the variables.

It makes it harder (in the current style) to add newtype instances.

Fix tests (by removing unnecessary type annotations). These type
annotations could have done the intended thing, but only with scoped
type variables (which was not used). Easier to remove them.